Working on otp and RR controllers, need to test jwt, edits on firstform, made genres/services db

+++ Project/Managers/Implementations/RatingAndReviewManager.cs
@@ -18,7 +18,7 @@ namespace Managers.Implementations
         }
 
         // Checks valid review fields
-        public bool CheckReviewFields(string titleSelected, int uRating, string uReview)
+        public bool CheckReviewFields(string titleSelected, float uRating, string uReview)
         {
 
             const int MAX_REVIEW_CHARACTERS = 2500;
@@ -43,7 +43,7 @@ namespace Managers.Implementations
 
         }
 
-        public bool SubmitReviewRating(string dispName, string titleSelected, int uRating, string uReview)
+        public bool SubmitReviewRating(string dispName, string titleSelected, float uRating, string uReview)
         {
             try
             {
@@ -76,7 +76,7 @@ namespace Managers.Implementations
             }
         }
 
-        public bool UpdateReviewRating(string dispName, string titleSelected, int uRating, string uReview)
+        public bool UpdateReviewRating(string dispName, string titleSelected, float uRating, string uReview)
         {
             RatingAndReview updateReview = new RatingAndReview(dispName, titleSelected, uRating, uReview);
             bool isValidReview = CheckReviewFields(titleSelected, uRating, uReview);
